refactor: bookmark to notification scope
